requirements.txt
telethon
gspread
oauth2client
python-dotenv
pytz
google-auth
google-auth-oauthlib
google-auth-httplib2
pytesseract
Pillow
APScheduler
